refactor(proxy): change rebuild_request return type from int to ssize_t
Updated rebuild_request() to return ssize_t instead of int, allowing it to
propagate negative error codes and correctly represent buffer sizes.

+++ include/common/rebuild_request.h
@@ -15,7 +15,7 @@
  * @return 0 on success & -1 on failure
  */
 
-int rebuild_request(HttpRequest *req, char *buffer,char *client_ip ,size_t buffer_size);
+ssize_t rebuild_request(HttpRequest *req, char *buffer,char *client_ip ,size_t buffer_size);
 
 /**
  * @brief Function to get client IP from socket
